requirements.txt
aiohttp==3.9.0
python-socketio==5.10.0
